Part 3: Minimum and Maximum Longitude and Latitudes + GUI Overhaul
A single (1) display that combines all input streams, and only outputs GPS events in the defined range of Latitude and Longitude.

Labels show the current setting of the restriction.
The GUI element displays the data in an identical form to the display field in (2), as long as the data is in the range.